fix(restructure): updated components & heatmap
